GRIDEDIT 265: Fix type conversions (#47)
This fixes a bunch of sonarcloud code smells and
gets rid of the remaining boost warnings.

+++ include/MeshKernel/Smoother.hpp
@@ -53,7 +53,7 @@ namespace meshkernel
         /// @brief node
         /// @brief connectedNode
         /// @returns
-        [[nodiscard]] auto GetWeight(int node, int connectedNode)
+        [[nodiscard]] auto GetWeight(size_t node, int connectedNode)
         {
             return m_weights[node][connectedNode];
         }
@@ -62,7 +62,7 @@ namespace meshkernel
         /// @brief node
         /// @brief connectedNode
         /// @returns
-        [[nodiscard]] auto GetConnectedNodeIndex(int node, int connectedNode)
+        [[nodiscard]] auto GetConnectedNodeIndex(size_t node, int connectedNode)
         {
             return m_connectedNodes[node][connectedNode];
         }
@@ -70,7 +70,7 @@ namespace meshkernel
         /// @brief Get number of connected nodes
         /// @brief node
         /// @returns
-        [[nodiscard]] auto GetNumConnectedNodes(int node)
+        [[nodiscard]] auto GetNumConnectedNodes(size_t node)
         {
             return m_numConnectedNodes[node];
         }
@@ -96,23 +96,23 @@ namespace meshkernel
 
         /// Computes operators of the elliptic smoother by node (orthonet_comp_operators)
         /// @param[in] currentNode
-        void ComputeOperatorsNode(int currentNode);
+        void ComputeOperatorsNode(size_t currentNode);
 
         /// @brief Computes m_faceNodeMappingCache, m_sharedFacesCache, m_connectedNodes for the current node, required before computing xi and eta
         /// @param[in] currentNode
         /// @param[out] numSharedFaces
         /// @param[out] numConnectedNodes
-        void NodeAdministration(int currentNode,
-                                int& numSharedFaces,
-                                int& numConnectedNodes);
+        void NodeAdministration(size_t currentNode,
+                                size_t& numSharedFaces,
+                                size_t& numConnectedNodes);
 
         /// @brief Compute compute current node xi and eta (orthonet_assign_xieta)
         /// @param[in] currentNode
         /// @param[in] numSharedFaces
         /// @param[in] numConnectedNodes
-        void ComputeNodeXiEta(int currentNode,
-                              int numSharedFaces,
-                              int numConnectedNodes);
+        void ComputeNodeXiEta(size_t currentNode,
+                              size_t numSharedFaces,
+                              size_t numConnectedNodes);
 
         /// @brief Compute optimal edge angle
         /// @brief numFaceNodes
@@ -120,27 +120,27 @@ namespace meshkernel
         /// @brief theta2
         /// @brief isBoundaryEdge
         /// @returns
-        [[nodiscard]] double OptimalEdgeAngle(int numFaceNodes,
+        [[nodiscard]] double OptimalEdgeAngle(size_t numFaceNodes,
                                               double theta1 = -1.0,
                                               double theta2 = -1.0,
                                               bool isBoundaryEdge = false) const;
 
         /// @brief Allocate smoother operators
         /// @param[in] topologyIndex
-        void AllocateNodeOperators(int topologyIndex);
+        void AllocateNodeOperators(size_t topologyIndex);
 
         /// @brief If it is a new topology, save it
         /// @param[in] currentNode
         /// @param[in] numSharedFaces
         /// @param[in] numConnectedNodes
-        void SaveNodeTopologyIfNeeded(int currentNode,
-                                      int numSharedFaces,
-                                      int numConnectedNodes);
+        void SaveNodeTopologyIfNeeded(size_t currentNode,
+                                      size_t numSharedFaces,
+                                      size_t numConnectedNodes);
 
         /// @brief Computes local coordinates jacobian from the mapped jacobians m_Jxi and m_Jeta
         /// @param[in] currentNode
         /// @param[out] J
-        void ComputeJacobian(int currentNode, std::vector<double>& J) const;
+        void ComputeJacobian(size_t currentNode, std::vector<double>& J) const;
 
         /// @brief Compute the matrix norm
         /// @brief x
@@ -168,12 +168,12 @@ namespace meshkernel
         std::vector<std::vector<double>> m_ww2;               // weights
 
         // Smoother local caches
-        std::vector<int> m_sharedFacesCache;
+        std::vector<size_t> m_sharedFacesCache;
         std::vector<size_t> m_connectedNodesCache;
         std::vector<std::vector<size_t>> m_faceNodeMappingCache;
         std::vector<double> m_xiCache;
         std::vector<double> m_etaCache;
-        std::vector<int> m_boundaryEdgesCache;
+        std::vector<size_t> m_boundaryEdgesCache;
         std::vector<double> m_leftXFaceCenterCache;
         std::vector<double> m_leftYFaceCenterCache;
         std::vector<double> m_rightXFaceCenterCache;
@@ -182,22 +182,22 @@ namespace meshkernel
         std::vector<double> m_etasCache;
 
         // Smoother topologies
-        int m_numTopologies = 0;
-        std::vector<int> m_nodeTopologyMapping;
-        std::vector<int> m_numTopologyNodes;
-        std::vector<int> m_numTopologyFaces;
+        size_t m_numTopologies = 0;
+        std::vector<size_t> m_nodeTopologyMapping;
+        std::vector<size_t> m_numTopologyNodes;
+        std::vector<size_t> m_numTopologyFaces;
         std::vector<std::vector<double>> m_topologyXi;
         std::vector<std::vector<double>> m_topologyEta;
-        std::vector<std::vector<int>> m_topologySharedFaces;
+        std::vector<std::vector<size_t>> m_topologySharedFaces;
         std::vector<std::vector<std::vector<size_t>>> m_topologyFaceNodeMapping;
         std::vector<std::vector<size_t>> m_topologyConnectedNodes;
 
-        std::vector<int> m_numConnectedNodes;              // (nmk2)
+        std::vector<size_t> m_numConnectedNodes;           // (nmk2)
         std::vector<std::vector<size_t>> m_connectedNodes; // (kk2)
 
         // Class variables
-        int m_maximumNumConnectedNodes = 0;
-        int m_maximumNumSharedFaces = 0;
+        size_t m_maximumNumConnectedNodes = 0;
+        size_t m_maximumNumSharedFaces = 0;
 
         static constexpr int m_topologyInitialSize = 10;
         static constexpr double m_thetaTolerance = 1e-4;
